move types to references
