+++ src/index.ts
@@ -1,4 +1,4 @@
 export { RainbowProvider } from './RainbowProvider';
 export { handleProviderRequest } from './handleProviderRequest';
-export { IMessageSender } from './references';
-export { AddEthereumChainProposedChain } from './utils/chains';
+export { IMessageSender } from './references/messengers';
+export { AddEthereumChainProposedChain } from './references/chains';
