+++ src/references/appSession.ts
@@ -1,0 +1,3 @@
+import { Address } from 'viem';
+
+export type ActiveSession = { address: Address; chainId: number } | null;
