+++ src/references/messengers.ts
@@ -1,0 +1,105 @@
+import { RPCMethod } from './ethereum';
+
+export type RequestArguments = {
+  id?: number;
+  method: RPCMethod;
+  params?: Array<unknown>;
+};
+
+export type RequestResponse =
+  | {
+      id: number;
+      error: Error;
+      result?: never;
+    }
+  | {
+      id: number;
+      error?: never;
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      result: any;
+    };
+
+export interface Tab {
+  /**
+   * Optional.
+   * The title of the tab. This property is only present if the extension's manifest includes the "tabs" permission.
+   */
+  title?: string | undefined;
+  /**
+   * Optional.
+   * The ID of the tab. Tab IDs are unique within a browser session. Under some circumstances a Tab may not be assigned an ID, for example when querying foreign tabs using the sessions API, in which case a session ID may be present. Tab ID can also be set to chrome.tabs.TAB_ID_NONE for apps and devtools windows.
+   */
+  id?: number | undefined;
+}
+
+export interface IMessageSender {
+  /**
+   * The URL of the page or frame that opened the connection. If the sender is in an iframe, it will be iframe's URL not the URL of the page which hosts it.
+   * @since Chrome 28.
+   */
+  url?: string | undefined;
+  /** The tabs.Tab which opened the connection, if any. This property will only be present when the connection was opened from a tab (including content scripts), and only if the receiver is an extension, not an app. */
+  tab?: Tab | undefined;
+}
+
+export type CallbackOptions = {
+  /** The sender of the message. */
+  sender: IMessageSender;
+  /** The topic provided. */
+  topic: string;
+  /** An optional scoped identifier. */
+  id?: number | string;
+};
+
+type CallbackFunction<TPayload, TResponse> = (
+  payload: TPayload,
+  callbackOptions: CallbackOptions,
+) => Promise<TResponse>;
+
+export type ProviderRequestPayload = RequestArguments & {
+  id: number;
+  meta?: CallbackOptions;
+};
+
+type ProviderResponse = RequestResponse;
+
+export interface IProviderRequestTransport {
+  send(
+    payload: ProviderRequestPayload,
+    {
+      id,
+    }: {
+      id: number;
+    },
+  ): Promise<ProviderResponse>;
+  reply(
+    callback: (
+      payload: ProviderRequestPayload,
+      callbackOptions: CallbackOptions,
+    ) => Promise<ProviderResponse>,
+  ): Promise<void>;
+}
+
+export interface IMessenger {
+  /** Whether or not the messenger is available in the context. */
+  available: boolean;
+  /** Name of the messenger */
+  name: string;
+  /** Sends a message to the `reply` handler. */
+  send: <TPayload, TResponse>(
+    /** A scoped topic that the `reply` will listen for. */
+    topic: string,
+    /** The payload to send to the `reply` handler. */
+    payload: TPayload,
+    options?: {
+      /** Identify & scope the request via an ID. */
+      id?: string | number;
+    },
+  ) => Promise<TResponse>;
+  /** Replies to `send`. */
+  reply: <TPayload, TResponse>(
+    /** A scoped topic that was sent from `send`. */
+    topic: string,
+    callback: CallbackFunction<TPayload, TResponse>,
+  ) => () => void;
+}
